Cleanup log_deprecated move
