+++ rerun_py/rerun_sdk/rerun/log_deprecated/arrow.py
@@ -5,7 +5,7 @@ from typing import Any
 import numpy.typing as npt
 
 from rerun.log import Color
-from rerun.log.log_decorator import log_decorator
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
