+++ rerun_py/rerun_sdk/rerun/log_deprecated/camera.py
@@ -4,8 +4,8 @@ import numpy.typing as npt
 
 from rerun import bindings
 from rerun.components.pinhole import Pinhole, PinholeArray
-from rerun.log.error_utils import _send_warning
-from rerun.log.log_decorator import log_decorator
+from rerun.error_utils import _send_warning
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
