+++ rerun_py/rerun_sdk/rerun/log_deprecated/file.py
@@ -8,7 +8,7 @@ import numpy as np
 import numpy.typing as npt
 
 from rerun import bindings
-from rerun.log.log_decorator import log_decorator
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
