+++ rerun_py/rerun_sdk/rerun/log_deprecated/image.py
@@ -6,7 +6,7 @@ import numpy as np
 import numpy.typing as npt
 
 from rerun._rerun2.datatypes.tensor_data import TensorDataLike
-from rerun.log.log_decorator import log_decorator
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
