+++ rerun_py/rerun_sdk/rerun/log_deprecated/mesh.py
@@ -7,7 +7,7 @@ import numpy.typing as npt
 
 from rerun import bindings
 from rerun.log import Colors, _normalize_colors
-from rerun.log.log_decorator import log_decorator
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
