+++ rerun_py/rerun_sdk/rerun/log_deprecated/rects.py
@@ -7,8 +7,8 @@ import numpy.typing as npt
 
 from rerun.components.rect2d import RectFormat
 from rerun.log import Color, Colors, OptionalClassIds
-from rerun.log.error_utils import _send_warning
-from rerun.log.log_decorator import log_decorator
+from rerun.error_utils import _send_warning
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
