+++ rerun_py/rerun_sdk/rerun/log_deprecated/scalar.py
@@ -9,7 +9,7 @@ from rerun.components import instance_key_splat
 from rerun.components.scalar import ScalarArray, ScalarPlotPropsArray
 from rerun.log import Color, _normalize_colors
 from rerun.log.extension_components import _add_extension_components
-from rerun.log.log_decorator import log_decorator
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
