+++ rerun_py/rerun_sdk/rerun/log_deprecated/tensor.py
@@ -6,8 +6,8 @@ import numpy as np
 import numpy.typing as npt
 
 from rerun._rerun2.datatypes.tensor_data import TensorDataLike
-from rerun.log.error_utils import _send_warning
-from rerun.log.log_decorator import log_decorator
+from rerun.error_utils import _send_warning
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 __all__ = [
