+++ rerun_py/rerun_sdk/rerun/log_deprecated/text.py
@@ -4,7 +4,7 @@ import logging
 from typing import Any, Final
 
 from rerun.log import Color, _normalize_colors
-from rerun.log.log_decorator import log_decorator
+from rerun.log_deprecated.log_decorator import log_decorator
 from rerun.recording_stream import RecordingStream
 
 # Fully qualified to avoid circular import
